Refactor QPanelController and UI elements
- Adjusted calculations for Enclosure dimensions in QPanelController.cs to divide by 10.
- Introduced UpdateCubical method for handling cubical panel updates.
- Reduced button and grid sizes in ViewControls.xaml for improved UI layout.
- Changed Qty type from int to double in DesignItem.cs for precision.
- Added CubicalNumber property in Group.cs and improved SQL queries.
- Updated ItemType.cs to include a Type property for better classification.
- Added CalculationItem.cs to define a new class for calculation items.
- Updated ElectricalRange.csproj to version 1.1.6.0 and removed unused image resources.
- Enhanced SQL queries and error handling in SelectGroupViewModel.cs and ReferencesViewModel.cs.

+++ ElectricalRange/Data/Library/DesignItem.cs
@@ -17,7 +17,7 @@ namespace ProjectsNow.Data.Library
         public int LabelNumber { get; set; }
         public string Code { get; set; }
         public string Description { get; set; }
-        public int Qty { get; set; }
+        public double Qty { get; set; }
         public string ItemTable { get; set; }
         public string Note { get; set; }
